+++ src/ScottPlot5/ScottPlot5 Tests/Unit Tests/RenderTests/LegendTests.cs
@@ -290,6 +290,7 @@ internal class LegendTests
                 LineColor = Colors.Blue,
                 LineWidth = 3,
             };
+            plt.Legend.TightHorizontalWrapping = true;
             plt.Legend.ManualItems.Add(item);
         }