+++ src/ScottPlot5/ScottPlot5/Primitives/Legend.cs
@@ -53,6 +53,14 @@ public class Legend(Plot plot) : IPlottable, IHasOutline, IHasBackground, IHasSh
     public List<LegendItem> ManualItems { get; set; } = [];
 
     /// <summary>
+    /// If enabled, items in horizontal oriented legends will not
+    /// be aligned in columns but instead resized tightly to fit their contents
+    /// </summary>
+    public bool TightHorizontalWrapping { get; set; } = false;
+
+    public bool ShowItemRectangles_DEBUG { get; set; } = false;
+
+    /// <summary>
     /// Enabling this allows multi-language text in the figure legend,
     /// but may slow down the render loop.
     /// </summary>
@@ -264,8 +272,7 @@ public class Legend(Plot plot) : IPlottable, IHasOutline, IHasBackground, IHasSh
 
             item.LabelStyle.Render(canvas, labelRect.LeftCenter, paint, true);
 
-            bool SHOW_DEBUG_RECTANGLES = false;
-            if (SHOW_DEBUG_RECTANGLES)
+            if (ShowItemRectangles_DEBUG)
             {
                 Drawing.DrawRectangle(canvas, symbolRect, Colors.Magenta.WithAlpha(.2));
                 Drawing.DrawRectangle(canvas, labelRect, Colors.Magenta.WithAlpha(.2));